// lib/astro/jaimini/types.ts
//////////////////////////////////////////////////////////////
// NATIONPATH ASTRO ENGINE
// Jaimini Types
//////////////////////////////////////////////////////////////

export interface DarakarakaResult {
  planet: string;
  longitude: number;
}

export interface UpapadaResult {
  sign: string;
  house: number;
}

export interface ArudhaResult {
  house: number;
  sign: string;
}

export interface KarakamshaResult {
  sign: string;
}

export interface PadaResult {
  house: number;
}

export interface RashiDrishtiResult {
  from: string;
  to: string[];
}